+++ packages/components/dialog-tree-selector/index.ts
@@ -1,0 +1,9 @@
+import { withInstall } from '@composite-ware/utils'
+
+export * from './src'
+
+import DialogTreeSelector from './src/dialog-tree-selector.vue'
+
+export const CeDialogTreeSelector = withInstall(DialogTreeSelector)
+
+export default CeDialogTreeSelector