+++ packages/components/tree-selector/src/index.ts
@@ -2,7 +2,7 @@ import type { IDefineProps } from '../../types/index'
 import { withInstall } from '@composite-ware/utils'
 import TreeSelector from './tree-selector.vue'
 
-// import { } from './props'
+import { treeSelectotProp } from './props'
 
 export const CeTreeSelector = withInstall(TreeSelector)
-// export type ITitleProps = IDefineProps<typeof TitleProps>
+export type treeSelectotProps = IDefineProps<typeof treeSelectotProp>